skills/msgraph/requirements.txt
msgraph-sdk>=1.5.0
msgraph-beta-sdk>=1.5.0
msal>=1.28.0
azure-core>=1.30.0
